isIdCurrent {RMySQL}	R Documentation

Check if a database object is valid.

Description

Support function that verifies that an object holding a reference to a foreign object is still valid for communicating with the RDBMS. isIdCurrent will be deprecated in the near future; please use the dbIsValid() generic instead.

Usage

isIdCurrent(obj)

## S4 method for signature 'MySQLDriver'
dbIsValid(dbObj)

## S4 method for signature 'MySQLConnection'
dbIsValid(dbObj)

## S4 method for signature 'MySQLResult'
dbIsValid(dbObj)

Arguments

dbObj, obj
A MysqlDriver, MysqlConnection, MysqlResult.

Details

dbObjects are R/S-Plus remote references to foreign objects. This introduces differences to the object's semantics such as persistence (e.g., connections may be closed unexpectedly), thus this function provides a minimal verification to ensure that the foreign object being referenced can be contacted.

Value

a logical scalar.

Examples

dbIsValid(MySQL())

[Package RMySQL version 0.10.21 Index]