mysqlHasDefault {RMySQL}	R Documentation

Check if default database is available.

Description

RMySQL examples and tests connect to a database defined by the rs-dbi group in ~/.my.cnf. This function checks if that database is available, and if not, displays an informative message.

Usage

mysqlHasDefault()

Examples

if (mysqlHasDefault()) {
  db <- dbConnect(RMySQL::MySQL(), dbname = "test")
  dbListTables(db)
  dbDisconnect(db)
}

[Package RMySQL version 0.10.21 Index]